fix: address PR #2495 code quality issues
Schema validation gaps, dead code, duplicate logic, and CSS
duplication identified in the UI regression audit. Key changes:

- Add Vite commons stub plugin to fix fast build webview loading
- Add schema validation for OPEN_VSCODE_SETTINGS and SET_TAB messages
- Remove dead code (sumRunUsage, getSettingsViewProvider, SettingsViewEvents)
- Fix search state not clearing when history is cleared
- Extract tryRefreshClient helper to deduplicate retry logic
- Consolidate command constants with single source of truth from schemas
- Derive tab indices from SETTINGS_TAB_ORDER array
- Extract memory filesystem logic to settingsView/utils/memoryFileSystem.ts
- Deduplicate history/search CSS into shared/styles/historyStyles.ts
- Implement ClientRef pattern for explicit client reference mutation
- Update AGENTS.md to reference settingsView instead of deleted views
