+++ src/agent/core/flows/CycleServices.ts
@@ -81,3 +81,37 @@ type CycleParams = Record<string, unknown>;
 // Re-export for backward compatibility with existing node definitions
 export type ResponseCycleParams<_C = unknown> = CycleParams;
 export type ToolUseCycleParams<_C = unknown> = CycleParams;
+
+// ---------------------------------------------------------------------------
+// Client Ref Pattern for refreshClient
+// ---------------------------------------------------------------------------
+
+/**
+ * Mutable reference wrapper for client instances.
+ * Used by cycle flows to support token refresh during retries.
+ *
+ * The refreshClient callback updates `current` instead of mutating
+ * a closed-over variable, making the mutation explicit and type-safe.
+ */
+export interface ClientRef<C> {
+  current: C;
+}
+
+/**
+ * Creates a client ref and its associated refresh function.
+ * The refresh function updates the ref's `current` property with a new client.
+ *
+ * @param initialClient - The initial client instance
+ * @param getClient - Async function to obtain a fresh client (e.g., after token refresh)
+ * @returns Tuple of [clientRef, refreshClient function]
+ */
+export function createClientRef<C>(
+  initialClient: C,
+  getClient: () => Promise<C>,
+): [ClientRef<C>, () => Promise<void>] {
+  const ref: ClientRef<C> = { current: initialClient };
+  const refreshClient = async (): Promise<void> => {
+    ref.current = await getClient();
+  };
+  return [ref, refreshClient];
+}
