+++ src/commands/settings/index.ts
@@ -1,7 +1,6 @@
 export {
   registerSettingsViewCommands,
   initializeSettingsViewProvider,
-  getSettingsViewProvider,
   showSettingsView,
   settingsViewCommands,
 } from './settingsCommands';
